requirements.txt
opencv-python
numpy
ultralytics
streamlit
pillow